+++ src/utils/index.js
@@ -22,6 +22,7 @@ export { decodeBeeValue } from "./decodeBeeValue.js";
 export { ensurePathExists } from "./ensurePathExists.js";
 export { formatToBuffer } from "./formatToBuffer.js";
 export { formatToStr } from "./formatToStr.js";
+export { forwardEvent } from "./forwardEvent.js";
 export { generateSeed } from "./generateSeed.js";
 export { isPearRuntime } from "./isPearRuntime.js";
 export { normalizePath } from "./normalizePath.js";